Move processors to package
