--- conjure-python-core/src/main/java/com/palantir/conjure/python/FlatteningPackageNameProcessor.java
@@ -1,14 +1,0 @@
-/*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
- */
-
-package com.palantir.conjure.python;
-
-public final class FlatteningPackageNameProcessor implements PackageNameProcessor.Processor {
-
-    @Override
-    public String processPackageName(String packageName) {
-        return packageName.replace(".", "_");
-    }
-
-}
