+++ conjure-python-core/src/main/java/com/palantir/conjure/python/processors/PythonIdentifierSanitizer.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.conjure.python.poet;
+package com.palantir.conjure.python.processors;
 
 import com.google.common.collect.ImmutableSet;
 import java.util.Collections;
