+++ conjure-python-core/src/test/java/com/palantir/conjure/python/ConjurePythonGeneratorTest.java
@@ -19,8 +19,6 @@ package com.palantir.conjure.python;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.palantir.conjure.defs.Conjure;
-import com.palantir.conjure.python.client.ClientGenerator;
-import com.palantir.conjure.python.types.DefaultBeanGenerator;
 import com.palantir.conjure.spec.ConjureDefinition;
 import java.io.File;
 import java.io.IOException;
@@ -38,7 +36,6 @@ import org.junit.runner.RunWith;
 public final class ConjurePythonGeneratorTest {
 
     private final ConjurePythonGenerator generator = new ConjurePythonGenerator(
-            new DefaultBeanGenerator(), new ClientGenerator(),
             GeneratorConfiguration.builder()
                     .packageName("package-name")
                     .packageVersion("0.0.0")
